Project-03-AirCanvasUltra_Pro/requirements.txt
opencv-python
mediapipe
tensorflow
numpy
pillow